// src/shared/components/invoice-form/data-access/interfaces/invoice-form-details.interface.ts
export interface InvoiceFormDetailsInterface {
    id: number;
    numero_demande: string;
    operation: string;
    compte_client: string;
    nom_client: string;
    type_entreprise: string;
    justificatif: string | null;
    prix_unitaire: string;
    prix_ht: string;
    tva: string;
    prix_tva: string;
    prix_ttc: string;
    type_paiement: string;
    etat_paiement: string;
    date_soumission_facture: string | null;
    delai_ack: number;
    delai_traitement: number;
    delai_cloture: number;
    delai_escalade: number;
    description: string | null;
    cloture_a: string | null;
    nb_demande_soumises: number;
    nb_demande_traitees: number;
    nb_demande_cloturees: number;
    initie_par: number;
    recu_paiement: string | null;
    intervenant_id: number;
    traite_a: string | null;
    traite_par: number;
    cloture_par: number | null;
    abandonne_a: string | null;
    abandonne_par: number | null;
    acquitte_a: string | null;
    acquitte_par: number | null;
    commentaire_traitement: string | null;
    commentaire_cloture: string | null;
    nb_demande_identifiees: string;
    nb_demande_echouees: string;
    type: string;
    finalise_a: string | null;
    finalise_par: number | null;
    commentaire_finalisation: string | null;
    montant_formule: string | null;
    nb_cycle: number;
    nb_demande_crees: number;
    accepte: string;
    equipe_id: number | null;
    nb_sim_affectees: number;
    accepte_approbation: string;
    commentaire_approbation: string | null;
    approuve_par: string | null;
    approuve_a: string | null;
    source: string | null;
    recuperer_par: string | null;
    recuperer_a: string | null;
    commentaire_livraison: string | null;
    etat_soumission: string | null;
    etat_traitement: string | null;
    etat_finalisation: string | null;
    etat_cloture: string | null;
    statut: string;
    traitement: string | null;
    notation_cloture: string | null;
    created_at: string;
    updated_at: string;
    demandeur_nom: string;
    demandeur_prenoms: string;
    demandeur_contacts: string;
    demandeur_email: string;
    demandeur_matricule: string;
    prix_ttc_lettres: string | null;
    montant_total: string;
    impaye_anterieur: string;
    soldeDisponible: string;
    SoldeEngage: string;

    fournisseurInfo: FournisseurInfoInterface;
    clientInfo: ClientInfoInterface;
    tenant: TenantInfoInterface;
    intervenant: IndividualInterface;
    finalisateur: IndividualInterface | null;
    approbateur: IndividualInterface;
    facture: InvoiceInterface;
}

interface InvoiceInterface {
    id: number;
    reference: string;
    compte_client: string;
    nom_client: string;
    initie_par: number;
    traite_a: string;
    traite_par: number;
    approuve_par: number;
    approuve_a: string;
    numero_demande: string;
    operation: string;
    qte: number;
    prix_unitaire: string;
    prix_ht: string;
    tva: string;
    prix_tva: string;
    prix_ttc: string;
    statut: string;
    montant_paye: string;
    remise: string;
    date_limite: string | null;
    type_paiement: string;
    qrcode: string;
    created_at: string;
    updated_at: string;
}

export interface FournisseurInfoInterface {
    id: number;
    nom_entreprise: string;
    forme_juridique: string;
    pays: string;
    rccm: string;
    cc: string;
    regime_imposition: string | null;
    capital_social: string;
    acompte: string | null;
    solde: string | null;
    adresse: string;
    email: string;
    contact_1: string;
    contact_2: string;
    site_web: string;
    date_fondation: string | null;
    ordre: string;
    banque: string;
    centre_impot: string;
    compte: string;
    mobile_money: string | null;
    nom_responsable: string;
    contact_responsable: string;
    email_responsable: string;
    mode_paiement: string;
    logo: string;
    created_at: string | null;
    updated_at: string | null;
}

export interface ClientInfoInterface {
    id: number;
    tenant_code: string;
    nom_tenant: string;
    slug: string;
    domaine_activite: string;
    email_diffusion: string;
    email_diffusion_tenant: string;
    url_detection_appro: string | null;
    intervenant_id: number | null;
    nom_admin_tenant: string;
    contact_admin_tenant: string;
    email_admin_tenant: string;
    nom_gestionnaire: string;
    contact_gestionnnaire: string;
    email_gestionnnaire: string;
    gestionnaire_tenant_id: number | null;
    escalade_tenant_id: number | null;
    commercial_id: number | null;
    nom_commercial: string | null;
    contact_commercial: string | null;
    email_commercial: string | null;
    senior_manager_id: number | null;
    nom_senior_manager: string | null;
    contact_senior_manager: string | null;
    email_senior_manager: string | null;
    description: string;
    url_backend: string | null;
    url_minio: string | null;
    url_nodered: string | null;
    url_deploy_profil_supervision: string | null;
    url_frontend: string | null;
    url_demandes_sla: string | null;
    url_dossiers_demandes: string | null;
    url_systeme_paiement: string | null;
    ws_server: string | null;
    nb_max_users: number;
    lien_dashboard_grafana: string | null;
    lien_dashboard_appro: string | null;
    lien_dashboard_rejets: string | null;
    logo_tenant: string;
    application: string;
    suffixe_email: string;
    url_consommation: string | null;
    compte_client: string;
    segment_client: string | null;
    forme_juridique_code: string | null;
    regime_code: string | null;
    adresse: string;
    numero_rccm: string;
    forme_juridique: string;
    nom_gerant: string;
    fichier_rccm: string;
    numero_cc: string;
    regime: string;
    centre: string;
    fichier_dfe: string;
    admin_tenant_id: number | null;
    contact_gerant: string | null;
    email_gerant: string | null;
    piece_gerant: string | null;
    created_at: string;
    updated_at: string;
}

export interface TenantInfoInterface {
    id: number;
    transaction: string;
    numero_demande: string;
    nom_client: string;
    logo_client: string;
    slug: string;
    domaine_activite: string;
    email_diffusion: string | null;
    email_diffusion_client: string | null;
    url_detection_appro: string | null;
    nom_admin_client: string | null;
    contact_admin_client: string | null;
    email_admin_client: string | null;
    nom_gestionnaire: string | null;
    contact_gestionnaire: string | null;
    email_gestionnaire: string | null;
    gestionnaire_client_id: number | null;
    escalade_client_id: number | null;
    nom_commercial: string | null;
    contact_commercial: string | null;
    email_commercial: string | null;
    nom_senior_manager: string | null;
    contact_senior_manager: string | null;
    email_senior_manager: string | null;
    description: string | null;
    url_frontend: string | null;
    suffixe_email: string | null;
    compte_client: string;
    segment_client: string | null;
    forme_juridique_code: string | null;
    regime_code: string | null;
    adresse: string;
    numero_rccm: string;
    fichier_rccm: string;
    fichier_dfe: string;
    numero_cc: string;
    centre: string;
    nom_gerant: string;
    contact_gerant: string;
    email_gerant: string;
    piece_gerant: string;
    type_entreprise: string;
    traitement: string;
    statut: string;
    created_at: string;
    updated_at: string;
    username: string | null;
    password: string | null;
    lien_client: string | null;
    fichier_formation: string | null;
    code_client: string | null;
}

export interface IndividualInterface {
    id: number;
    matricule: string;
    nom: string;
    prenoms: string;
    username: string;
    adresse: string;
    contacts: string;
    email: string;
}